changed maven configuration, added assembly plugin packaging with dependencies, surefire plugin and heroku plugin
