+++ src/main/java/zpi/Main.java
@@ -8,11 +8,18 @@ import zpi.utils.Paths;
 public class Main {
 	public static void main(String[] args) {
 		Javalin app = Javalin.create(config -> {
-		}).start(7777);
-		
-		
+		}).start(getHerokuAssignedPort());
+
+
 		app.get(Paths.Web.INDEX, MainPageController.mainPage);
 		app.get(Paths.Web.SIMPLE_TAX, ComputeTaxController.computeTax);
-		
+	}
+
+	private static int getHerokuAssignedPort() {
+		String herokuPort = System.getenv("PORT");
+		if (herokuPort != null) {
+			return Integer.parseInt(herokuPort);
+		}
+		return 7000;
 	}
 }
