+++ src/test/java/TaxComputingTest.java
@@ -1,5 +1,3 @@
-package test.zpi.sales;
-
 import io.javalin.Javalin;
 import org.junit.Test;
